Product seller

+++ src/controllers/product.js
@@ -1,3 +1,5 @@
+/* eslint-disable consistent-return */
+/* eslint-disable max-len */
 const Sequelize = require('sequelize');
 const { cloudPathToFileName } = require('../helpers/converter');
 const { deleteImages } = require('../helpers/deleteArrayImages');
@@ -7,6 +9,7 @@ const Category = require('../models/category');
 const Product = require('../models/product');
 const ProductCategory = require('../models/productCategory');
 const ProductImage = require('../models/productImage');
+const User = require('../models/user');
 
 const { APP_URL } = process.env;
 
@@ -17,7 +20,7 @@ exports.getAllProduct = async (req, res) => {
   } = req.query;
   minPrice = parseInt(minPrice, 10) || 0;
   maxPrice = parseInt(maxPrice, 10) || 100000000;
-  limit = parseInt(limit, 10) || 20;
+  limit = parseInt(limit, 10) || 10;
   page = parseInt(page, 10) || 1;
   const dataName = ['search', 'minPrice', 'maxPrice'];
   const data = { search, minPrice, maxPrice };
@@ -31,8 +34,14 @@ exports.getAllProduct = async (req, res) => {
   const offset = (page - 1) * limit;
   const results = await Product.findAll({
     include: [
-      { model: ProductCategory },
-      { model: ProductImage },
+      {
+        model: ProductCategory,
+        attributes: ['id_category'],
+      },
+      {
+        model: ProductImage,
+        attributes: ['image'],
+      },
     ],
     where: {
       name: {
@@ -42,6 +51,9 @@ exports.getAllProduct = async (req, res) => {
         [Sequelize.Op.gte]: minPrice,
         [Sequelize.Op.lte]: maxPrice,
       },
+      stock: {
+        [Sequelize.Op.gte]: 1,
+      },
       is_deleted: 0,
     },
     limit,
@@ -49,6 +61,13 @@ exports.getAllProduct = async (req, res) => {
   });
   const count = await Product.count({
     where: {
+      name: {
+        [Sequelize.Op.like]: `%${search}%`,
+      },
+      price: {
+        [Sequelize.Op.gte]: minPrice,
+        [Sequelize.Op.lte]: maxPrice,
+      },
       is_deleted: 0,
     },
   });
@@ -63,13 +82,112 @@ exports.getAllProduct = async (req, res) => {
   return responseHandler(res, 200, 'List of products', results, pageInfo);
 };
 
+exports.getProductBySeller = async (req, res) => {
+  try {
+    const id = req.params.seller_id;
+    const seller = await User.findAll({
+      where: {
+        id,
+        id_role: 2,
+      },
+    });
+    if (!seller || seller.length < 1) {
+      return responseHandler(res, 404, 'Seller not found');
+    }
+    const { search = '' } = req.query;
+    let {
+      minPrice, maxPrice, limit, page,
+    } = req.query;
+    minPrice = parseInt(minPrice, 10) || 0;
+    maxPrice = parseInt(maxPrice, 10) || 100000000;
+    limit = parseInt(limit, 10) || 10;
+    page = parseInt(page, 10) || 1;
+    const dataName = ['search', 'minPrice', 'maxPrice'];
+    const data = { search, minPrice, maxPrice };
+    let url = `${APP_URL}/product/seller/${id}?`;
+    dataName.forEach((x) => {
+      if (req.query[x]) {
+        data[x] = req.query[x];
+        url = `${url}${x}=${data[x]}&`;
+      }
+    });
+    const offset = (page - 1) * limit;
+    const product = await Product.findAll({
+      include: [
+        {
+          model: ProductCategory,
+          attributes: ['id_category'],
+        },
+        {
+          model: ProductImage,
+          attributes: ['image'],
+        },
+      ],
+      where: {
+        name: {
+          [Sequelize.Op.like]: `%${search}%`,
+        },
+        price: {
+          [Sequelize.Op.gte]: minPrice,
+          [Sequelize.Op.lte]: maxPrice,
+        },
+        stock: {
+          [Sequelize.Op.gte]: 1,
+        },
+        seller_id: id,
+        is_deleted: 0,
+      },
+      limit,
+      offset,
+    });
+    const count = await Product.count({
+      where: {
+        name: {
+          [Sequelize.Op.like]: `%${search}%`,
+        },
+        price: {
+          [Sequelize.Op.gte]: minPrice,
+          [Sequelize.Op.lte]: maxPrice,
+        },
+        seller_id: id,
+        is_deleted: 0,
+      },
+    });
+    const last = Math.ceil(count / limit);
+    const pageInfo = {
+      prev: page > 1 ? `${url}page=${page - 1}&limit=${limit}` : null,
+      next: page < last ? `${url}page=${page + 1}&limit=${limit}` : null,
+      totalData: count,
+      currentPage: page,
+      lastPage: last,
+    };
+    if (product.length > 0) {
+      return responseHandler(res, 200, 'Product list', product, pageInfo);
+    }
+    return responseHandler(res, 200, 'Seller has no product yet', null, null);
+  } catch {
+    return responseHandler(res, 500, 'Unexpected error');
+  }
+};
+
 exports.createProduct = async (req, res) => {
   try {
+    if (req.user.role !== 2) {
+      if (req.files) {
+        deleteImages(req.files);
+      }
+      return responseHandler(res, 401, 'You are not able to do this action');
+    }
     const listIdCategory = req.body.id_category.split(',');
     if (listIdCategory.length < 1) {
       return responseHandler(res, 400, 'Please enter at least 1 category', null, null);
     }
-    const product = await Product.create(req.body);
+    const dataProduct = {};
+    Object.keys(req.body).forEach((x) => {
+      dataProduct[x] = req.body[x];
+    });
+    dataProduct.seller_id = req.user.id;
+    const product = await Product.create(dataProduct);
     const data = { id_product: product.dataValues.id };
     if (req.files) {
       req.files.forEach(async (pic) => {
@@ -108,12 +226,31 @@ exports.createProduct = async (req, res) => {
 };
 
 exports.productDetail = async (req, res) => {
-  const { id } = req.params;
-  const product = await Product.findByPk(id);
-  if (product && product.is_deleted === false) {
-    return responseHandler(res, 200, 'Product detail', product, null);
+  try {
+    const { id } = req.params;
+    const product = await Product.findAll({
+      include: [
+        {
+          model: ProductCategory,
+          attributes: ['id_category'],
+        },
+        {
+          model: ProductImage,
+          attributes: ['image'],
+        },
+      ],
+      where: {
+        id,
+        is_deleted: 0,
+      },
+    });
+    if (product.length > 0) {
+      return responseHandler(res, 200, 'Product detail', product, null);
+    }
+    return responseHandler(res, 404, 'Product not found', null, null);
+  } catch {
+    return responseHandler(res, 500, 'Unexpected error');
   }
-  return responseHandler(res, 404, 'Product not found', null, null);
 };
 
 exports.updateProduct = async (req, res) => {
@@ -126,9 +263,16 @@ exports.updateProduct = async (req, res) => {
       }
       return responseHandler(res, 404, 'Product not found', null, null);
     }
+    if (req.user.role !== 2 && req.user.id !== product.dataValues.seller_id) {
+      if (req.files) {
+        deleteImages(req.files);
+      }
+      return responseHandler(res, 401, 'You are not able to do this action');
+    }
     Object.keys(req.body).forEach((data) => {
       product[data] = req.body[data];
     });
+    product.seller_id = req.user.id;
     await product.save();
     const productImage = await ProductImage.findAll({
       where: {
@@ -159,23 +303,33 @@ exports.updateProduct = async (req, res) => {
 };
 
 exports.deleteProduct = async (req, res) => {
-  const { id } = req.params;
-  const product = await Product.findByPk(id);
-  if (product && product.dataValues.is_deleted === false) {
-    product.is_deleted = 1;
-    await product.save();
-    const productImage = await ProductImage.findAll({
-      where: {
-        id_product: id,
-      },
-    });
-    if (productImage) {
-      productImage.map(async (data) => {
-        await data.destroy();
-        deleteFile(cloudPathToFileName(data.image));
-      });
+  try {
+    const { id } = req.params;
+    const product = await Product.findByPk(id);
+    if (req.user.id !== product.dataValues.seller_id) {
+      return responseHandler(res, 401, 'You are not able to do this action');
     }
-    return responseHandler(res, 200, 'Product was deleted', null, null);
+    if (product && product.dataValues.is_deleted === false) {
+      product.is_deleted = 1;
+      await product.save();
+      const productImage = await ProductImage.findAll({
+        where: {
+          id_product: id,
+        },
+      });
+      if (productImage) {
+        productImage.map(async (data) => {
+          await data.destroy();
+          deleteFile(cloudPathToFileName(data.image));
+        });
+      }
+      return responseHandler(res, 200, 'Product was deleted', null, null);
+    }
+    return responseHandler(res, 404, 'Product not found', null, null);
+  } catch (e) {
+    if (req.files) {
+      deleteImages(req.files);
+    }
+    return responseHandler(res, 400, 'Can\'t delete', e, null);
   }
-  return responseHandler(res, 404, 'Product not found', null, null);
 };
